Move `ProtosForTest.java` to unique package.
