+++ modules/session/src/test/java/com/nautilus_technologies/tsubakuro/impl/low/sql/SessionLinkImplTest.java
@@ -15,8 +15,8 @@ import com.nautilus_technologies.tsubakuro.channel.common.sql.ResponseWireHandle
 import com.nautilus_technologies.tsubakuro.protos.Distiller;
 import com.nautilus_technologies.tsubakuro.protos.RequestProtos;
 import com.nautilus_technologies.tsubakuro.protos.ResponseProtos;
-import com.nautilus_technologies.tsubakuro.protos.ProtosForTest;
-    
+import com.nautilus_technologies.tsubakuro.session.ProtosForTest;
+
 import org.junit.jupiter.api.Test;
 
 class SessionLinkImplTest {
